[Insights][Bug] Filter out excluded officers in action strategies endpoint logic (Recidiviz/recidiviz-data#37634)
## Description of the change

In the action strategies endpoint, filters the officer list for officers
where `include_in_outcomes` is true.

The actual fix is very small, but I also added a couple tests related to
this situation in `outlier_routes_test`.

The action strategy endpoint was trying to pull outcomes for all
officers, rather than just the officers with outcomes, which was causing
a action strategies error message in prod (see ticket for specific
example). I had forgotten to make the filtering change here during the
relevant refactor work.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes
[#recidiviz-dashboards/7192](https://github.com/Recidiviz/recidiviz-dashboards/issues/7192)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 095af2263cae3124a14fff14b6394eac9a165df6