[US_AZ] 100 Days past DTP ACIS date, date becomes invalid (Recidiviz/recidiviz-data#37661)
## Description of the change

We made a change to TPR a while back that made it so that folks who
passed their ACIS date by 100 days were no longer considered eligible.
For these folks, they had likely been released on CSBD/ERCD but kept
their ACIS date in the system. For this PR I made the equal change to
the DTP opportunity.

In order to clean things up, I ported the majority of the code to the
query fragments file for easy reuse.

Loaded to sandbox under `shalin`. Seeing 53 folks who are **no longer
eligible** for the ACIS_DTP opportunity as a result of this change.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#37253

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2ae89d6c988a5bb6c3cda46bed19ca650308a4b8