[Aggregated Metrics] Migrate impact_reports dataset to use new optimized query format (Recidiviz/recidiviz-data#37655)
## Description of the change

Updates the `impact_reports` dataset to use the new optimized metrics
query format to generate metrics.

This PR generates two sets of aggregated metrics, both which get written
to the `impact_reports` dataset:
* `usage` - Monthly, weekly and yearly metrics related to distinct
active users of various sorts
* `impact_funnel` - Daily metrics related to our impact funnel (who is
almost eligible, eligible, viewed and eligible, etc)

The previous `impact_reports` views did a lot of data math shenanigans
to generate metric rows ONLY for the exact week and month periods we
care about for the reports, but the way the app script is set up, it
doesn't really matter if we have extra rows. For weekly and monthly
metrics, we now just generate a metric period for every day in the
lookback period, whether or not it's properly aligned with the start of
the month. This isn't a huge deal now because the lookback is still
relatively small and these queries are much less expensive.

This PR also hooks these new metrics tables into the impact reports app
script code. I loaded the changes to a sandbox and generated a test
report based off of my changes, which I compared to a test report built
off of deployed code.

* Test report for TN using deployed views / code:
[link](https://docs.google.com/document/d/1K8sZHinFYHRPcw4HsnIUAuVBEfSeBqIAVrVh3mOVHR4/edit?tab=t.0#heading=h.7tap3soi8gte)
* Test report for TN using sandbox views / code from this PR:
[link](https://docs.google.com/document/d/1uPYYNqbiNQJ0yJIDJ4lnQUOoltDt18u1XvYYz_mDi8k/edit?tab=t.0).
* I changed the title of this document but otherwise made no other
changes and as far as I can tell the outputs are identical.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35911

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1eec0ff43942c909b604fbb57c4fc93448a399b2