[Tools] Fixes bug in load_views_to_sandbox script when newly added views are ignored (Recidiviz/recidiviz-data#37780)
## Description of the change

We used to have a check in `load_views_to_sandbox` that would identify
when there are newly added views (not yet deployed to BQ) that we are
excluding from a view sandbox which have children that we ARE including
in the sandbox. Previously we were just failing the whole script when we
identified these views, but it really seems more user friendly to just
include them if we find them.

When `reference_views.us_or_location_metadata` (a US_OR-specific view)
was added in a recent commit and someone else tried to load a
US_IX-specific sandbox that loaded `reference_views.location_metadata`,
we would automatically try to ignore that OR-specific view, then fail
because we could not load `location_metadata`.

This PR updates the sandbox script to include that view anyway, even if
the arguments suggest that it should be excluded.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

See issue discussed here:
https://recidiviz.slack.com/archives/C01J3JV7962/p1737998300447419

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e239cd232de53b6f7028c3c839869a026c289576