[Testing] Deprecate legacy ENUM_RAW_TEXT fixture type and path (Recidiviz/recidiviz-data#37773)
## Description of the change

Removes `DirectIngestFixtureDataFileType.ENUM_RAW_TEXT`, as these
fixtures will not be managed by Janet. We now have a function that takes
in a `StateCode` and file tag to load the fixtures.

Shout out to the ~8k lines of unused fixtures, lol

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| **Type: Non-breaking refactor** | change addresses some tech debt item
or prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Supports Recidiviz/recidiviz-data#36159

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 090bc504a5d7cad3d99026400d80c6104bd58a89